// apps/web/src/generated/models/dto-series-episode-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * commentPlayer API
 * Watch commentPlayer API documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { DtoSeriesEpisodeWatchHistoryItem } from './dto-series-episode-watch-history-item';

export interface DtoSeriesEpisodeResponse {
    /**
     * エピソード番号
     */
    'episode'?: number;
    /**
     * ファイル名
     */
    'file_name'?: string;
    /**
     * エピソードのサブタイトル
     */
    'subtitle'?: string;
    /**
     * ビデオID
     */
    'video_id'?: number;
    /**
     * 動画の総再生数
     */
    'views'?: number;
    /**
     * 視聴履歴
     */
    'watch_history'?: Array<DtoSeriesEpisodeWatchHistoryItem>;
}